refactor: 바우처 타입 패키지 구조 변경
