+++ src/main/java/org/prgrms/kdt/domain/voucher/controller/VoucherController.java
@@ -3,9 +3,9 @@ package org.prgrms.kdt.domain.voucher.controller;
 import org.prgrms.kdt.domain.command.types.CommandType;
 import org.prgrms.kdt.domain.customer.model.Customer;
 import org.prgrms.kdt.domain.customer.service.CustomerService;
+import org.prgrms.kdt.domain.voucher.model.VoucherType;
 import org.prgrms.kdt.util.Console;
 import org.prgrms.kdt.domain.voucher.model.Voucher;
-import org.prgrms.kdt.domain.voucher.types.VoucherType;
 import org.prgrms.kdt.domain.voucher.service.VoucherService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
