+++ src/main/java/org/prgrms/kdt/domain/voucher/model/VoucherType.java
@@ -1,4 +1,4 @@
-package org.prgrms.kdt.domain.voucher.types;
+package org.prgrms.kdt.domain.voucher.model;
 
 import java.util.Arrays;
 
@@ -13,7 +13,7 @@ public enum VoucherType {
     }
 
     public static VoucherType findVoucherType(String type){
-        return Arrays.stream(values())
+        return Arrays.stream(VoucherType.values())
                 .filter(value -> value.type.equals(type))
                 .findAny()
                 .orElseThrow(() -> new IllegalArgumentException("지원하지 않는 바우처타입 입니다."));
