+++ src/test/java/org/prgrms/kdt/domain/voucher/service/VoucherServiceTest.java
@@ -9,8 +9,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.prgrms.kdt.domain.voucher.model.FixedAmountVoucher;
 import org.prgrms.kdt.domain.voucher.model.PercentDiscountVoucher;
 import org.prgrms.kdt.domain.voucher.model.Voucher;
+import org.prgrms.kdt.domain.voucher.model.VoucherType;
 import org.prgrms.kdt.domain.voucher.repository.VoucherRepository;
-import org.prgrms.kdt.domain.voucher.types.VoucherType;
 
 import java.util.Arrays;
 import java.util.List;
